update to symlink for copying to var/spacedev

+++ .devcontainer/features/spacefx-dev/updateContent.sh
@@ -13,6 +13,15 @@ source /etc/os-release
 
 source /spacefx-dev/.env
 
-ln -s ${SPACEFX_DIR_FOR_HOST} ${SPACEFX_DIR}
+if [[ ! -d "${SPACEFX_DIR_FOR_HOST}" ]]; then
+    mkdir -p ${SPACEFX_DIR_FOR_HOST}
+fi
 
+# Setup a symlink between the devcontainer SPACEFX_DIR to the host SPACEFX_DIR
+# if one doesn't already exist
+if [[ ! -L "${SPACEFX_DIR}" ]]; then
+    ln -s ${SPACEFX_DIR_FOR_HOST} ${SPACEFX_DIR}
+fi
+
+# Copy all the files from the devcontainer to the host
 cp /azure-orbital-space-sdk-setup/* ${SPACEFX_DIR_FOR_HOST} -r
